requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
requests
beautifulsoup4
lxml
confluent-kafka
psycopg2-binary
faker
pyspark